remove support for legacy dat format
