+++ data/regressions.tsv
@@ -12,4 +12,3 @@ https://googleads.g.doubleclick.net/pagead/id	https://m.youtube.com/	xhr	true
 https://yt3.ggpht.com/-6Yjb9ya2ETs/AAAAAAAAAAI/AAAAAAAAAAA/jRCIdTp_htQ/s68-c-k-no-mo-rj-c0xffffff/photo.jpg	https://m.youtube.com/	image	false
 https://m.youtube.com/yts/jsbin/player-plasma-ias-phone-en_GB-vfl-TJfdP/base.js	https://m.youtube.com/	script	false
 https://m.youtube.com/yts/cssbin/mobile-polymer-player-ias-mweb-2x-vflsTwlMM.css	https://m.youtube.com/	stylesheet	false
-https://m.youtube.com/csi_204?v=2&s=youtube_mobile&action=home&yt_sts=n&st=169&vpni=3&vpil=3&yt_lt=cold&is_nav=1&csn=uM9fXYLTLtKQ1waPx6KoDQ&p=h2&t=tcp&yt_vis=1&pis=downloading&yt_fn=what_to_watch&GetBrowse_rid=0x76c7c55799a38bb5&c=MWEB&cver=2.20190822.03.01&yt_li=0&rt=ai.289,apr.292,srt.78,fpt.261,nreqs.3,nress.78,nrese.201,wffs.245,wffe.275,pdr.300,ips.323,ipc.338,cpt.339,vpl.431,aft.339,ps.464	https://m.youtube.com/	image	true
