+++ src/data_format/mod.rs
@@ -5,10 +5,9 @@
 //! serialization/deserialization implementations and can automatically dispatch to the appropriate
 //! one.
 
-mod legacy;
 mod v0;
 
-pub mod utils;
+pub(crate) mod utils;
 
 use crate::blocker::Blocker;
 use crate::cosmetic_filter_cache::CosmeticFilterCache;
@@ -23,31 +22,12 @@ const ADBLOCK_RUST_DAT_MAGIC: [u8; 4] = [0xd1, 0xd9, 0x3a, 0xaf];
 /// Note that this does not implement `Serialize` directly, as it is composed of parts which must
 /// be serialized independently. Instead, use the `serialize` method.
 pub(crate) enum SerializeFormat<'a> {
-    Legacy(legacy::SerializeFormat<'a>),
     V0(v0::SerializeFormat<'a>),
 }
 
 #[derive(Debug)]
 pub enum SerializationError {
     RmpSerdeError(rmp_serde::encode::Error),
-    GzError(std::io::Error),
-}
-
-/// Since two different versions of `rmp-serde` are being used, errors must be converted to a
-/// single implementation.
-impl From<rmp_serde_legacy::encode::Error> for SerializationError {
-    fn from(e: rmp_serde_legacy::encode::Error) -> Self {
-        use rmp_serde::encode::Error as EncodeError;
-        use rmp_serde_legacy::encode::Error as LegacyEncodeError;
-
-        let new_error = match e {
-            LegacyEncodeError::InvalidValueWrite(e) => EncodeError::InvalidValueWrite(e),
-            LegacyEncodeError::UnknownLength => EncodeError::UnknownLength,
-            LegacyEncodeError::DepthLimitExceeded => EncodeError::DepthLimitExceeded,
-            LegacyEncodeError::Syntax(e) => EncodeError::Syntax(e),
-        };
-        Self::RmpSerdeError(new_error)
-    }
 }
 
 impl From<rmp_serde::encode::Error> for SerializationError {
@@ -56,24 +36,13 @@ impl From<rmp_serde::encode::Error> for SerializationError {
     }
 }
 
-impl From<std::io::Error> for SerializationError {
-    fn from(e: std::io::Error) -> Self {
-        Self::GzError(e)
-    }
-}
-
 impl<'a> SerializeFormat<'a> {
-    pub(crate) fn build(blocker: &'a Blocker, cfc: &'a CosmeticFilterCache, legacy: bool) -> Self {
-        if legacy {
-            Self::Legacy(legacy::SerializeFormat::from((blocker, cfc)))
-        } else {
-            Self::V0(v0::SerializeFormat::from((blocker, cfc)))
-        }
+    pub(crate) fn build(blocker: &'a Blocker, cfc: &'a CosmeticFilterCache) -> Self {
+        Self::V0(v0::SerializeFormat::from((blocker, cfc)))
     }
 
     pub(crate) fn serialize(&self) -> Result<Vec<u8>, SerializationError> {
         match self {
-            Self::Legacy(v) => v.serialize(),
             Self::V0(v) => v.serialize(),
         }
     }
@@ -85,7 +54,6 @@ impl<'a> SerializeFormat<'a> {
 /// Note that this does not implement `Deserialize` directly, as it is composed of parts which must
 /// be deserialized independently. Instead, use the `deserialize` method.
 pub(crate) enum DeserializeFormat {
-    Legacy(legacy::DeserializeFormat),
     V0(v0::DeserializeFormat),
 }
 
@@ -94,28 +62,10 @@ pub enum DeserializationError {
     RmpSerdeError(rmp_serde::decode::Error),
     UnsupportedFormatVersion(u8),
     NoHeaderFound,
-}
-
-/// Since two different versions of `rmp-serde` are being used, errors must be converted to a
-/// single implementation.
-impl From<rmp_serde_legacy::decode::Error> for DeserializationError {
-    fn from(e: rmp_serde_legacy::decode::Error) -> Self {
-        use rmp_serde::decode::Error as DecodeError;
-        use rmp_serde_legacy::decode::Error as LegacyDecodeError;
-
-        let new_error = match e {
-            LegacyDecodeError::InvalidMarkerRead(e) => DecodeError::InvalidMarkerRead(e),
-            LegacyDecodeError::InvalidDataRead(e) => DecodeError::InvalidDataRead(e),
-            LegacyDecodeError::TypeMismatch(m) => DecodeError::TypeMismatch(m),
-            LegacyDecodeError::OutOfRange => DecodeError::OutOfRange,
-            LegacyDecodeError::LengthMismatch(l) => DecodeError::LengthMismatch(l),
-            LegacyDecodeError::Uncategorized(e) => DecodeError::Uncategorized(e),
-            LegacyDecodeError::Syntax(e) => DecodeError::Syntax(e),
-            LegacyDecodeError::Utf8Error(e) => DecodeError::Utf8Error(e),
-            LegacyDecodeError::DepthLimitExceeded => DecodeError::DepthLimitExceeded,
-        };
-        Self::RmpSerdeError(new_error)
-    }
+    /// Support for the legacy gzip-compressed data format was removed in version 0.8.0 of this
+    /// crate. If you still need it for some reason, you can convert it using 0.7.x by
+    /// deserializing and then reserializing it into the newer V0 format.
+    LegacyFormatNoLongerSupported,
 }
 
 impl From<rmp_serde::decode::Error> for DeserializationError {
@@ -127,26 +77,23 @@ impl From<rmp_serde::decode::Error> for DeserializationError {
 impl DeserializeFormat {
     pub(crate) fn build(self) -> (Blocker, CosmeticFilterCache) {
         match self {
-            Self::Legacy(v) => v.into(),
             Self::V0(v) => v.into(),
         }
     }
 
     pub(crate) fn deserialize(serialized: &[u8]) -> Result<Self, DeserializationError> {
-        /// adblock-rust has always used flate2 1.0.x for the legacy format, which has never
-        /// changed the header sequence from these 10 bits when the GzEncoder is left uncustomized.
+        /// adblock-rust's legacy DAT format has always used flate2 1.0.x, which has never changed
+        /// the header sequence from these 10 bits when the GzEncoder is left uncustomized.
         const FLATE2_GZ_HEADER_BYTES: [u8; 10] = [31, 139, 8, 0, 0, 0, 0, 0, 0, 255];
 
-        if serialized.starts_with(&FLATE2_GZ_HEADER_BYTES) {
-            Ok(Self::Legacy(legacy::DeserializeFormat::deserialize(
-                serialized,
-            )?))
-        } else if serialized.starts_with(&ADBLOCK_RUST_DAT_MAGIC) {
+        if serialized.starts_with(&ADBLOCK_RUST_DAT_MAGIC) {
             let version = serialized[ADBLOCK_RUST_DAT_MAGIC.len()];
             match version {
                 0 => Ok(Self::V0(v0::DeserializeFormat::deserialize(serialized)?)),
                 v => Err(DeserializationError::UnsupportedFormatVersion(v)),
             }
+        } else if serialized.starts_with(&FLATE2_GZ_HEADER_BYTES) {
+            Err(DeserializationError::LegacyFormatNoLongerSupported)
         } else {
             Err(DeserializationError::NoHeaderFound)
         }
