+++ src/engine.rs
@@ -115,34 +115,15 @@ impl Engine {
     pub fn serialize_raw(&self) -> Result<Vec<u8>, BlockerError> {
         use crate::data_format::SerializeFormat;
 
-        let serialize_format = SerializeFormat::build(&self.blocker, &self.cosmetic_cache, false);
+        let serialize_format = SerializeFormat::build(&self.blocker, &self.cosmetic_cache);
 
         serialize_format.serialize().map_err(|_e| {
             BlockerError::SerializationError
         })
     }
 
-    /// Serializes the `Engine` into a compressed binary format so that it can be quickly reloaded later.
-    ///
-    /// The data format generated from this method is _not_ just a gzip compressed version of
-    /// `serialize_raw`; it is a distinct format. If you'd like to convert data between the two
-    /// formats, deserialize it into the `Engine` first, then serialize the appropriate type.
-    ///
-    /// This method will be removed in a future release. Going forwards, if you'd like to use a
-    /// compressed binary format, use `serialize_raw` and bring your own compression/decompression.
-    pub fn serialize_compressed(&self) -> Result<Vec<u8>, BlockerError> {
-        use crate::data_format::SerializeFormat;
-
-        let serialize_format = SerializeFormat::build(&self.blocker, &self.cosmetic_cache, true);
-
-        serialize_format.serialize().map_err(|_e| {
-            BlockerError::SerializationError
-        })
-    }
-
-    /// Deserialize the `Engine` from the binary format generated by `Engine::serialize_compressed`
-    /// or `Engine::serialize_raw`. The method will automatically select the correct
-    /// deserialization implementation.
+    /// Deserialize the `Engine` from the binary format generated by `Engine::serialize_raw`. The
+    /// method will automatically select the correct deserialization implementation.
     pub fn deserialize(&mut self, serialized: &[u8]) -> Result<(), BlockerError> {
         use crate::data_format::DeserializeFormat;
         let current_tags = self.blocker.tags_enabled();
@@ -500,7 +481,7 @@ mod tests {
         let mut engine = Engine::from_rules(&[
             "ad-banner",
         ], Default::default());
-        let serialized = engine.serialize_compressed().unwrap();
+        let serialized = engine.serialize_raw().unwrap();
         println!("Engine serialized: {:?}", serialized);
         engine.deserialize(&serialized).unwrap();
     }
@@ -511,7 +492,7 @@ mod tests {
             "ad-banner$tag=abc",
         ], Default::default());
         engine.use_tags(&["abc"]);
-        let serialized = engine.serialize_compressed().unwrap();
+        let serialized = engine.serialize_raw().unwrap();
         println!("Engine serialized: {:?}", serialized);
         engine.deserialize(&serialized).unwrap();
     }
@@ -527,7 +508,7 @@ mod tests {
             Resource::simple("noopcss", MimeType::TextCss, ""),
         ]);
 
-        let serialized = engine.serialize_compressed().unwrap();
+        let serialized = engine.serialize_raw().unwrap();
         println!("Engine serialized: {:?}", serialized);
         engine.deserialize(&serialized).unwrap();
     }
