+++ tests/live.rs
@@ -198,7 +198,7 @@ fn get_blocker_engine_deserialized() -> Engine {
     let brave_service_key = std::env::var("BRAVE_SERVICE_KEY")
         .expect("Must set the $BRAVE_SERVICE_KEY environment variable to execute live tests.");
 
-    let dat_url = "https://adblock-data.s3.brave.com/4/rs-ABPFilterParserData.dat";
+    let dat_url = "https://adblock-data.s3.brave.com/ios/latest.dat";
     let download_client = reqwest::Client::new();
     let resp_bytes_fut = download_client
         .get(dat_url)
